spreads render content across files
