+++ public/js/views/index.js
@@ -1,7 +1,8 @@
 import app from '../app.js';
 import renderAsideGroups from './renderAside/renderAsideGroups.js'
 import renderNavBar from './renderNavbar.js';
-import { renderContent, renderContentGroups, renderContentTasks, renderModalButton, renderFooter, renderTaskDetailsPopup} from './everythingElse.temp.js';
+import renderContent from './content/renderContent.js';
+import {  renderModalButton, renderFooter, renderTaskDetailsPopup} from './everythingElse.temp.js';
 let viewIndex = 1;
 export default class View {
   // What the app looks like, what the user can see and do, User Interface
@@ -26,8 +27,6 @@ export default class View {
     renderAsideGroups(title);
     renderNavBar();
     renderContent();
-    renderContentGroups();
-    renderContentTasks();
     renderModalButton();
     renderFooter();
     renderTaskDetailsPopup(app.tasks[0]);
